requirements.txt
wtforms==3.0.1
flask==2.3.2
flask_wtf==1.1.1
flask_login==0.6.2
flask_sqlalchemy==3.0.3
email_validator==2.0.0.post2
Jinja2==3.1.2
Werkzeug==2.3.4